Enumerate all currently attached RTLSDRs. (02)
For this we use rtlsdr_get_device_usb_strings, passing it the
index of the device we want to use and a buffer to hold its
USB VID:PID pair and serial number.

Don't let the escape sequence spook you, it's pretty simple once
you break it down: remember \t is a tab character.